usecases/mwaa-dag-factory-example/requirements/requirements.txt
apache-airflow-providers-amazon==9.15.0
dag-factory==1.0.1